convert all uint to int in entities
ref #202

+++ src/FilterLists.Data/EntityTypeConfigurations/FilterListTypeConfiguration.cs
@@ -31,7 +31,7 @@ namespace FilterLists.Data.EntityTypeConfigurations
             entityTypeBuilder.Property(x => x.IssuesUrl)
                              .HasColumnType("TEXT");
             entityTypeBuilder.Property(x => x.LicenseId)
-                             .HasDefaultValue((uint)5)
+                             .HasDefaultValue(5)
                              .IsRequired();
             entityTypeBuilder.Property(x => x.Name)
                              .HasColumnType("VARCHAR(126)")
